modififier node finish

+++ nodes/object_data/ObjectDataNode.py
@@ -20,7 +20,7 @@ class RSNodeObjectDataNode(RenderStackNode):
     float_value: FloatProperty(name='Value', update=update_node)
     string_value: StringProperty(name='Value', update=update_node)
     bool_value: BoolProperty(name='On', update=update_node)
-    int_value: FloatProperty(name='Value', update=update_node)
+    int_value: IntProperty(name='Value', update=update_node)
 
     color_value: FloatVectorProperty(name='Color', update=update_node, subtype='COLOR',
                                      default=(1.0, 1.0, 1.0),
